Fix Ruby Swaparoo private repo hack for new bundler versions
Newer bundler versions were giving an error on `method(:gem)`. It seems
to be now named `_gem` in newer versions (not sure how the 'gem' method
is still working, possibly that is now a built-in which smartly defers
to bundler?) Anyway, this fixes it.

There is at least one arguably better way to do private repos in docker
now (I prefer using a container to forward SSH creds)