Use environment variable after parsing from args
Running the gitpull procedure appears to define a local variable with a
default value that is meant to be overridden in args. In a scenario
where the environment is not provided by the arguments, the procedures
attempt to run with the `args.environment` value, not the local
`environment` variable.

This ends up crashing the program if the environment is not part of the
args.

Fixes #163